KaustavGPT-Chatbot/requirements.txt
# ── Core LangChain / LangGraph ────────────────────────────────────────────────
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0

# ── LLM providers ────────────────────────────────────────────────────────────
langchain-groq>=0.2.0

# ── Embeddings ────────────────────────────────────────────────────────────────
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0

# ── Vector store ──────────────────────────────────────────────────────────────
faiss-cpu>=1.8.0

# ── Document loaders ──────────────────────────────────────────────────────────
pypdf>=4.0.0

# ── Tools / Search ────────────────────────────────────────────────────────────
duckduckgo-search>=6.0.0
requests>=2.31.0

# ── Checkpointer ──────────────────────────────────────────────────────────────
# SQLiteSaver is bundled with langgraph; no separate install needed

# ── Frontend ──────────────────────────────────────────────────────────────────
streamlit>=1.38.0

# ── Env / Utils ───────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# If need use pip install -U ddgs